3-ая итерация сдачи. Исправил все замечания после ревью. Переработал логику метода updateItem в классе ItemServiceImpl, вынес часть логирования в сервис, добавил исключение для работы с проверкой владельца, убрал параметр userId в методах findByText и findById интерфейса ItemRepository.

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -1,7 +1,9 @@
 package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.item.exception.NotItemOwnerException;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.mapper.ItemMapper;
 import ru.practicum.shareit.item.model.Item;
@@ -11,6 +13,7 @@ import ru.practicum.shareit.user.repository.UserRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ItemServiceImpl implements ItemService {
@@ -20,32 +23,35 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public ItemDto addNewItem(long userId, ItemDto itemDto) {
+        log.info("Обработка запроса на добавление нового предмета.");
         userRepository.findByUserId(userId);
         itemDto.setOwner(userId);
         Item item = ItemMapper.mapToItem(itemDto);
         Item createdItem = itemRepository.create(userId, item);
+        log.info("Предмет item = {}, успешно создан", createdItem.toString());
         return ItemMapper.mapToItemDto(createdItem);
     }
 
     @Override
     public ItemDto updateItem(long userId, long itemId, ItemDto itemDto) {
-        Item item = new Item();
-        item.setName(itemDto.getName());
-        item.setDescription(itemDto.getDescription());
-        item.setAvailable(itemDto.getAvailable());
-
+        log.info("Запрос на обновление данных вещи с id = {} пользователем с id = {}.", itemId, userId);
+        itemRepository.findById(itemId);
+        checkItemOwner(userId, itemId);
+        Item item = ItemMapper.mapToItem(itemDto);
         Item updatedItem = itemRepository.update(userId, itemId, item);
         return ItemMapper.mapToItemDto(updatedItem);
     }
 
     @Override
     public ItemDto getItemById(long userId, long itemId) {
-        Item item = itemRepository.findById(userId, itemId);
+        log.info("Запрос на получение вещи с id = {} пользователем с id = {}", itemId, userId);
+        Item item = itemRepository.findById(itemId);
         return ItemMapper.mapToItemDto(item);
     }
 
     @Override
     public List<ItemDto> getUsersItems(long userId) {
+        log.info("Запрос на получение всех вещей пользователя с id = {}", userId);
         return itemRepository.findUsersItems(userId).stream()
                 .map(ItemMapper::mapToItemDto)
                 .collect(Collectors.toList());
@@ -53,13 +59,28 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public List<ItemDto> getItemsByText(long userId, String text) {
-        return itemRepository.findByText(userId, text).stream()
+        log.info("Запрос на получение доступных вещей по тексту = '{}'", text);
+        if (text == null || text.isBlank()) {
+            log.warn("Поисковый текст пуст.");
+            return List.of();
+        }
+        return itemRepository.findByText(text).stream()
                 .map(ItemMapper::mapToItemDto)
                 .collect(Collectors.toList());
     }
 
     @Override
     public void deleteItem(long userId, long itemId) {
+        log.info("Запрос на удаление вещи с id = {} пользователем с id = {}", itemId, userId);
+        checkItemOwner(userId, itemId);
         itemRepository.deleteByItemId(userId, itemId);
     }
+
+    private void checkItemOwner(long userId, long itemId) {
+        if (itemRepository.findById(itemId).getOwner() != userId) {
+            log.warn("Пользователь с id = {} не является владельцем вещи с id = {}.", userId, itemId);
+            throw new NotItemOwnerException("Вы не являетесь владельцем этой вещи.");
+        }
+    }
 }
+
